move getVideoCourse and getUserData to utils
