+++ src/pages/myCourses/classRoom/videoCourseRoom/[courseId].tsx
@@ -1,13 +1,12 @@
 import React, { useState, useEffect, useRef, useContext } from "react";
 import Image from "next/image";
-import { doc, getDoc } from "firebase/firestore";
 import { useRouter } from "next/router";
 import styled from "styled-components";
 import Swal from "sweetalert2";
 import parse from "html-react-parser";
 import { FullScreen, useFullScreenHandle } from "react-full-screen";
 import Head from "next/head";
-import { db } from "../../../../../lib/firebase";
+import { getUserData, getVideoCourse } from "../../../../utils/firestore";
 import { AuthContext } from "../../../../contexts/authContext";
 import Play from "../../../../../public/play.png";
 import Pause from "../../../../../public/pause.png";
@@ -895,8 +894,7 @@ function VideoRoom({ courseId, courseData }: { courseId: string; courseData: Cou
   useEffect(() => {
     const getBoughtCourses = async () => {
       if (!userData.uid) return;
-      const userRef = doc(db, "users", userData.uid);
-      const docSnap = await getDoc(userRef);
+      const docSnap = await getUserData(userData.uid);
       if (docSnap.exists()) {
         const courses = docSnap.data().boughtCourses as string[];
         setBoughtCourses(courses);
@@ -911,8 +909,7 @@ function VideoRoom({ courseId, courseData }: { courseId: string; courseData: Cou
 
   useEffect(() => {
     const getCourseInfo = async () => {
-      const teacherRef = doc(db, "users", courseData.teacherId);
-      const teacherSnap = await getDoc(teacherRef);
+      const teacherSnap = await getUserData(courseData.teacherId);
       if (teacherSnap.exists()) {
         const teacherName = teacherSnap.data().username as string;
         const teacherAvatar = teacherSnap.data().photoURL as string;
@@ -923,8 +920,7 @@ function VideoRoom({ courseId, courseData }: { courseId: string; courseData: Cou
 
       if (!Array.isArray(courseData.reviews)) return;
       courseData.reviews.forEach(async (review: { comments: string; score: number; userId: string }, index) => {
-        const userRef = doc(db, "users", review.userId);
-        const userSnap = await getDoc(userRef);
+        const userSnap = await getUserData(review.userId);
         if (!userSnap.exists()) return;
         const username = userSnap.data().username as string;
         const avatar = userSnap.data().photoURL as string;
@@ -1021,23 +1017,8 @@ function VideoRoom({ courseId, courseData }: { courseId: string; courseData: Cou
 export default VideoRoom;
 
 export async function getServerSideProps({ params }: { params: { courseId: string } }) {
-  const docRef = doc(db, "video_courses", params.courseId);
-  const docSnap = await getDoc(docRef);
-  if (!docSnap.exists()) {
-    return {
-      notFound: true,
-    };
-  }
-  const id = docSnap.data().id as string;
-  const name = docSnap.data().name as string;
-  const chapters = docSnap.data().chapters as ChapterInterface[];
-  const introduction = docSnap.data().introduction as string;
-  const introductionVideo = docSnap.data().introductionVideo as string;
-  const teacherId = docSnap.data().teacher_id as string;
-  const cover = docSnap.data().cover as string;
-  const price = docSnap.data().price as number;
-  const reviews = docSnap.data().reviews as ReviewInterface[];
-  const courseData = { id, name, chapters, introduction, introductionVideo, teacherId, cover, price, reviews };
+  const { courseId } = params;
+  const courseData = await getVideoCourse(courseId);
 
-  return { props: { courseId: params.courseId, courseData } };
+  return { props: { courseId, courseData } };
 }
