+++ src/pages/videoCourses/courseDetail/[courseId].tsx
@@ -1,13 +1,14 @@
 import React, { useState, useEffect, useRef, useContext } from "react";
 import Image from "next/image";
 import Head from "next/head";
-import { doc, getDoc, updateDoc, arrayUnion } from "firebase/firestore";
+import { doc, updateDoc, arrayUnion } from "firebase/firestore";
 import { useRouter } from "next/router";
 import styled from "styled-components";
 import Swal from "sweetalert2";
 import parse from "html-react-parser";
 import { SetterOrUpdater, useRecoilState } from "recoil";
 import { FullScreen, useFullScreenHandle } from "react-full-screen";
+import { getVideoCourse, getUserData } from "../../../utils/firestore";
 import { db } from "../../../../lib/firebase";
 import { AuthContext } from "../../../contexts/authContext";
 import { orderQtyState, showMemberModalState } from "../../../utils/recoil";
@@ -915,8 +916,7 @@ function CourseInfo({ courseId, courseData }: { courseId: string; courseData: Co
   useEffect(() => {
     const getBoughtCourses = async () => {
       if (!userData.uid) return;
-      const userRef = doc(db, "users", userData.uid);
-      const docSnap = await getDoc(userRef);
+      const docSnap = await getUserData(userData.uid);
       if (docSnap.exists()) {
         const courses = docSnap.data().boughtCourses as string[];
         setBoughtCourses(courses);
@@ -927,8 +927,7 @@ function CourseInfo({ courseId, courseData }: { courseId: string; courseData: Co
 
   useEffect(() => {
     const getCourseInfo = async () => {
-      const teacherRef = doc(db, "users", courseData.teacherId);
-      const teacherSnap = await getDoc(teacherRef);
+      const teacherSnap = await getUserData(courseData.teacherId);
       if (teacherSnap.exists()) {
         const teacherName = teacherSnap.data().username as string;
         const teacherAvatar = teacherSnap.data().photoURL as string;
@@ -939,8 +938,7 @@ function CourseInfo({ courseId, courseData }: { courseId: string; courseData: Co
 
       if (!Array.isArray(courseData.reviews)) return;
       courseData.reviews.forEach(async (review: { comments: string; score: number; userId: string }, index) => {
-        const userRef = doc(db, "users", review.userId);
-        const userSnap = await getDoc(userRef);
+        const userSnap = await getUserData(review.userId);
         if (!userSnap.exists()) return;
         const username = userSnap.data().username as string;
         const avatar = userSnap.data().photoURL as string;
@@ -967,7 +965,7 @@ function CourseInfo({ courseId, courseData }: { courseId: string; courseData: Co
       }),
     });
     Swal.fire({ title: "已加入購物車！", confirmButtonColor: "#5d7262", icon: "success" });
-    const docSnap = await getDoc(userRef);
+    const docSnap = await getUserData(userData.uid);
     if (docSnap.exists()) {
       const cartItems = docSnap.data().cartItems as [];
       const qty = cartItems.length;
@@ -1051,23 +1049,7 @@ function CourseInfo({ courseId, courseData }: { courseId: string; courseData: Co
 export default CourseInfo;
 
 export async function getServerSideProps({ params }: { params: { courseId: string } }) {
-  const docRef = doc(db, "video_courses", params.courseId);
-  const docSnap = await getDoc(docRef);
-  if (!docSnap.exists()) {
-    return {
-      notFound: true,
-    };
-  }
-  const id = docSnap.data().id as string;
-  const name = docSnap.data().name as string;
-  const chapters = docSnap.data().chapters as ChapterInterface[];
-  const introduction = docSnap.data().introduction as string;
-  const introductionVideo = docSnap.data().introductionVideo as string;
-  const teacherId = docSnap.data().teacher_id as string;
-  const cover = docSnap.data().cover as string;
-  const price = docSnap.data().price as number;
-  const reviews = docSnap.data().reviews as ReviewInterface[];
-  const courseData = { id, name, chapters, introduction, introductionVideo, teacherId, cover, price, reviews };
-
-  return { props: { courseId: params.courseId, courseData } };
+  const { courseId } = params;
+  const courseData = await getVideoCourse(courseId);
+  return { props: { courseId, courseData } };
 }
